global: archivers refactoring
* Refactors the archivers.

* Adds bagit metadata jsonschema generation.

+++ invenio_sipstore/archivers/base_archiver.py
@@ -95,25 +95,28 @@ class BaseArchiver(object):
         self.path = folder
         self._create_directories(folder)
 
-    def create(self, filesdir="", metadatadir="", sipfiles=None):
+    def create(self, filesdir="", metadatadir="", sipfiles=None,
+               dry_run=False):
         """Create the archive.
 
-        :param str filesdir: the directory where to put files if necessary
-        :param str metadatadir: the directory where to put metadata if
-            necessary
-        :param sipfiles: list of SIPFile objects to write. By default
+        :param str filesdir: directory to which the data files will be written.
+        :param str metadatadir: directory to which the metadata files will be
+            written.
+        :param sipfiles: a list of SIPFile objects to write. By default
             it's all of the SIPFiles attached to the SIP.
-        :returns: a dictionnary with the filenames as keys, and size and
-            checksum as value
-        :rtype: dict
+        :returns: a list with the dictionaries consiting of 'filename', 'size',
+            'checksum' and the 'path', which contains the absolute path on
+            the filesystem to which the file was written. Files from SIPFiles
+            contain also 'file_uuid' key, which is the UUID (primary key) of
+            the related FileInstance object.
         """
         sipfiles = sipfiles or self.sip.files
-        files_info = self._copy_files(sipfiles, filesdir)
+        files_info = self._copy_files(sipfiles, filesdir, dry_run=dry_run)
         metadata = self.get_metadata()
         for filename, content in metadata.items():
-            files_info.update(self._save_file(
+            files_info.append(self._save_file(
                 join(metadatadir, filename),
-                content))
+                content, dry_run=dry_run))
         return files_info
 
     def finalize(self):
@@ -132,7 +135,7 @@ class BaseArchiver(object):
         if not self.fs.exists(dirname):
             self.fs.makedir(dirname, recursive=True)
 
-    def _copy_files(self, files, filesdir=""):
+    def _copy_files(self, files, filesdir="", dry_run=False):
         """Copy the files inside the new storage.
 
         Takes care of adding self.path at the beginning.
@@ -140,11 +143,17 @@ class BaseArchiver(object):
         :param files: the list of files to copy
         :type files: list(:py:class:`invenio_sipstore.models.SIPFile`)
         :param str filesdir: the directory where to copy files
-        :returns: a dictionnary with the filenames as keys, and size and
-            checksum as value
-        :rtype: dict
+        :param bool dry_run: When True, the files will not be copied,
+            and only the file information will be returned. No changes will be
+            made to disk or the database.
+        :returns: a list with the dictionaries consiting of 'filename', 'size',
+            'checksum' and the 'path', which contains the absolute path on
+            the filesystem to which the file was written. Files from SIPFiles
+            contain also 'file_uuid' key, which is the UUID (primary key) of
+            the related FileInstance object.
+        :rtype: list(dict)
         """
-        result = {}
+        result = []
 
         total_size = sum([file.size for file in files])
         copied_size = 0
@@ -153,15 +162,16 @@ class BaseArchiver(object):
             src_fs, path = opener.parse(file.storage_location)
             filename = join(filesdir, file.filepath)
             calculated_filename = join(self.path, filename)
-            self._create_directories(dirname(calculated_filename))
-            copyfile(src_fs, path,
-                     self.fs, calculated_filename)
-            result[filename] = {
+            if not dry_run:
+                self._create_directories(dirname(calculated_filename))
+                copyfile(src_fs, path, self.fs, calculated_filename)
+            result.append({
+                'filename': filename,
                 'size': file.size,
                 'checksum': file.checksum,
                 'path': self.fs.getsyspath(calculated_filename),
                 'file_uuid': str(file.file.id)
-            }
+            })
             copied_size += file.size
 
             sipstore_archiver_status.send({
@@ -174,15 +184,19 @@ class BaseArchiver(object):
             })
         return result
 
-    def _save_file(self, filename, content):
+    def _save_file(self, filename, content, dry_run=False):
         """Save the content inside the file.
 
         Takes care of adding self.path in the filename.
 
         :param str filename: the name of the file
         :param str content: the content of the file
-        :returns: a dictionnary with the filenames as keys, and size and
-            checksum as value
+        :param bool dry_run: When True, the files will not be copied,
+            and only the file information will be returned. No changes will be
+            made to disk or the database.
+        :returns: a dictionary describing the saved file, containing 'filename'
+            'size', 'checksum' and 'path' which contains the absolute path on
+            the filesystem to which the file was written.
         :rtype: dict
 
         ..warning::
@@ -192,15 +206,14 @@ class BaseArchiver(object):
         """
         content = b(content)
         filenamepath = join(self.path, filename)
-        self._create_directories(dirname(filenamepath))
-
-        with self.fs.open(filenamepath, 'wb') as fp:
-            fp.write(content)
+        if not dry_run:
+            self._create_directories(dirname(filenamepath))
+            with self.fs.open(filenamepath, 'wb') as fp:
+                fp.write(content)
 
         return {
-            filename: {
-                'size': len(content),
-                'checksum': 'md5:' + str(md5(content).hexdigest()),
-                'path': self.fs.getsyspath(filenamepath)
-            }
+            'filename': filename,
+            'size': len(content),
+            'checksum': 'md5:' + str(md5(content).hexdigest()),
+            'path': self.fs.getsyspath(filenamepath)
         }
